requirements.txt
python-decouple==3.1
requests==2.20.0
psycopg2-binary==2.7.4
APScheduler==3.5.1
